treewide: Move constant attributes from hl to core dialect.
