+++ include/vast/Dialect/HighLevel/HighLevelAttributes.hpp
@@ -2,32 +2,9 @@
 
 #pragma once
 
-#include "vast/Util/Warnings.hpp"
-
-VAST_RELAX_WARNINGS
-#include <llvm/ADT/APSInt.h>
-
-#include <llvm/Support/Locale.h>
-VAST_UNRELAX_WARNINGS
-
-#include "vast/Util/Common.hpp"
-
 #include "vast/Dialect/HighLevel/HighLevelDialect.hpp"
 #include "vast/Interfaces/TypeQualifiersInterfaces.hpp"
-
-#include "mlir/IR/BuiltinAttributes.h"
 
 #define GET_ATTRDEF_CLASSES
 #include "vast/Dialect/HighLevel/HighLevelAttributes.h.inc"
 
-#include "vast/Util/TypeList.hpp"
-
-namespace vast::hl {
-
-    using high_level_typed_attrs = util::type_list<
-        BooleanAttr, IntegerAttr, FloatAttr, StringAttr, StringLiteralAttr, VoidAttr
-    >;
-
-    bool is_high_level_typed_attr(mlir::Attribute);
-
-} // namespace vast::hl
